Moved rest of `sa_insights` iteration into `try` block
The `get_sa_insights` function had the core logic split between
the `try` block and the function code itself.  If the `try` was
successful, then the rest of the function operated as expected.

When the `except` was encountered, it would move to the function
code and fail because the `sa_insights` reference doesn't exist
in any scope.

The rest of the function logic was moved into the `try` block.

Additionally, a default `except` was added that will print the
exception encountered to stdout.

The function continuing to iterate through the list of projects
was preserved in this patch.